requirements.txt
anthropic>=0.25.0
openai>=1.12.0
requests>=2.31.0
python-dotenv>=1.0.0
Pillow>=10.0.0
tweepy>=4.14.0
PyNaCl>=1.5.0
